:rocket: Move Preprocessor rules
