+++ src/app/lib/utilities/preprocessor/rules/filterEmptyGroupItems/index.js
