+++ src/app/lib/utilities/preprocessor/rules/filterGroupsWithoutStraplines/index.js
